refactor(acor): extract V1 trie operations to trie.go
